Adding support for Proguard files.
These files typically have the extension .pro which are currently identified as Prolog files. This commit modifies languages.yml to add Proguard (and remove .pro as one of the additional extensions for Prolog), adds some samples for Proguard, and renames a few of the prolog samples from *.pl to *.pro to support the old use case.