Added IDL to list of languages to be recognised.
IDL (Interactive Data Language; http://www.exelisvis.com/ProductsServices/IDL.aspx) is
a language for scientific computing which is used a lot within the fields of physics,
astronomy and satellite imaging. Previously IDL-based projects on Github were being
recognised as Prolog, as IDL files have the same file extension (.pro) as one of the
alternative extensions for Prolog files.

IDL is listed as a supported language on the Pygments web page
(http://pygments.org/languages/) under the name 'IDL', so I assume that is the correct
lexer name. Please let me know if this is wrong!

I have placed a few samples of IDL code in the correct samples directory - it is
*very* different to prolog, so the Bayesian classifier should be able to separate them
fairly easily.