Don't suppress Gemfile.lock changes
Changes to this file are important and should be reviewed as a matter of
course when reviewing pull requests. While `Gemfile.lock` is generated,
in part, from `Gemfile`, the `Gemfile.lock` file can (and frequently
does) change irrespective of `Gemfile` changes.

With this file suppressed it's too easy to miss otherwise unwanted or
even dangerous dependency changes in reviews. Suppressing the file
encourages it to be ignored and it most certainly should not be.