Add support for G-code language
This is a special language controlling 3D printers (by RepRap, Makerbot,
Ultimaker etc.). It is not a general purpose programming language, but
still contains commands for e.g. looping. On the other hand, most of the
time it will be generated by another program, not hand-written. Hence I
classified it as "data".

Specification:
* http://reprap.org/wiki/G-code

Some repositories with examples:
* https://github.com/reprappro/Mendel
* https://github.com/BLLIP/bllip-parser
* https://github.com/MakerGear/M2