Invert conditions for matching a heuristic
Previously, a heuristic had to match all the possible language
candidates to be used, and could include other languages that are not
candidates. This inverts the logic so a heuristic only has to include a
subset of the candidates to be considered a match. This change makes it
so existing heuristics don’t break when new conflicting language
extensions are added.