Adds support for Line of Code to the linguist executable:
We wanted a quick way to get LOC/language for a repo, and linguist
had all the tools, but didn't offer it as a CLI flag. So, I added
it. :)

I did make an attempt at refactoring the linguist executable a bit,
added very simple presenters that can write to any IO object.

I hope you'll find this useful. Incidentally, it's mentioned as
a want in [Issue 3131](https://github.com/github/linguist/issues/3131)