added ".def" file extension entry for Modula-2
Modula-2 libraries consist of a definition module (file extension .def) and an implementation module (file extension .mod). This is comparable to .h and .c files in C and C++. Linguist not recognising .def, but .mod alone for Modula-2 would be like not recognising .c for C and C++. It is broken without.