add support for Meta Programming System (MPS)
Meta Programming System is a open source tool from JetBrains:

https://www.jetbrains.com/mps/

It is used to define custom (programming) languages and code generators
and it also used to use these languages. MPS stores the programs in its
AST form serialised to XML. MPS by default uses two file extensions:
most commonly .mps:

https://github.com/search?p=57&q=extension%3Amps+NOT+nothack&type=Code&utf8=✓

and far less common .mpsr

https://github.com/search?utf8=✓&q=extension%3Ampsr+NOT+nothack&type=Code

The later is used when programms are stored over mutiple files the first
is used when a complete programm is stored in a single file.

Of the 43k files all XML files I checked are indeed MPS programms. There
are at least two repositories using the file extension that are not MPS
programms but some other text data most probably some for of map data
and one repository that uses it for binary files.