Avoid loading samples when Classifier isn't used
samples.json is a very large (3.7M) json file, which allocates 183209 ruby
strings, which takes 8.2MB of memory according to memsize_of.

Instead of keeping the cache around when loading language.rb, we can
instead just load the JSON and allow it to be GC'd after we use it.
Cache will still be used if the Classifier is invoked.