Fix line splitting for POSIX files
When working on files with complete lines (always terminating with a
newline character) `BlobHelper` would produce an extra empty field in
the lines array.

This has the side effect of increasing the reported line count on every
file which adheres to this convention.

File data is now passed through `chomp` before splitting to remove any
possible final newline marker. This has no effects on files with
incomplete lines and files with one or more empty lines at their end.

Resolves #4563.