Make shebang detection cheaper
String#lines can potentially allocate *many* strings depending on the
number of newlines in the string.  That might be OK if the file we're
looking at is actually a file with a shebang in it (because we might
eventually use every line).  But it is not OK when the file isn't a
shebang file because we end up allocating many objects for no reason.

This commit reduces the number of allocations it takes to determine
whether or not the string is interesting to the Shebang class.

Here is a benchmark:

```ruby
require "linguist/shebang"

def allocs
  x = GC.stat(:total_allocated_objects)
  yield
  GC.stat(:total_allocated_objects) - x
end

file = "foo\n" * 1000

p allocs {
  Linguist::Shebang.interpreter(file)
}
```

Before this commit:

```
$ bundle _1.17.2_ exec ruby -I lib:test thing.rb
1003
```

After this commit:

```
$ bundle _1.17.2_ exec ruby -I lib:test thing.rb
2
```