Improve C++ language color
Minor changes to the C++ language color to make it match with the official colors of Google's Fuchsia OS (#ed1d7d) and the Zircon Microkernel. This change does not make it hard to distinguish or make it hard to recognize for all users that are used to the previous color. The change does not introduce any new problems. The two colors look nearly identical.