Improve C++ color
Minor change to the C++ language color from #f34b7d to #ed1d7d. The new color matches with the official colors of Google's Fuchsia OS. It's brighter and more vibrant. This change does not make it hard to recognize for all users that are used to the color since the two colors are almost indistinguishable.