Replace OpenEdge ABL grammar
The original source is no longer maintained and the maintainer has now deleted the repository. This replacement uses the grammar that is currently the base for the VSCode, Atom, Sublime and TextMate. A slight tweak is needed in the grammar compiler as the grammar is not stored in the path we expect.