Improve punctuation tokenization and word split
The tokenizer previously ignored most (but not all) punctuation
characters. Now all ASCII printable characters are included in tokens.

These characters are often very salient features when detecting the
programming language. In some cases, even more than identifiers.

A sequence of consecutive non-alphanumeric characters will be recognized
as a single token (e.g. == != >= <==>) with the following exceptions:

- Comment tokenization behavior is preserved, so `/*` or `<!--` are still
ignored as part of comments. As a side-effect, some edge cases are now
fixed (e.g. `<!---->`, `<!--Comment-->`).

- SGML tokens are mow more restrictive, so `<a>`, `<?xml>`, `<!DOCTYPE>` are still SGML recognized, but `<!--` or `<--`
are not. This also simplifies the corresponding lexer action.

- Parenthesis (`()`), square brackets (`[]`) and curly brackets (`{}`)
are tokenized independently.

Word splitting is also altered by this change:

- `[.@#/*]` are not considered part of an identifier anymore. But a single
character from [.@#] is recognized as part of a word if it's the first
one. So `.globl` in Assembly is still a single token,
but `console.log` is now split as `console` and `.log`. In Objective-C,
`@interface` will still be a single token, but `mail@example.com` will
be 3 tokens. This change leads to less unique tokens in the vocabulary,
which helps classification too.