Update languages.yml
Adding DataFlex from DataAccess. I picked the id 424 because it was the next one available, but I suspect that it might be a good idea to put a higher number there not to conflict with other reacent suggestions. I have not created any highlight rules. But I might consider spending time on that if this proposal get accepted.