Map .qmd files to markdown syntax highlighting
The quarto markdown system (derivative of pandoc, see https://quarto.org/) uses .qmd files for markdown. This is analagous to how R Markdown uses .rmd -- these files are run for computations and produce an .md file.

There is already an RMarkdown language mode in languages.yml which I could alternately add this to, but since that one is essentially also aliasing markdown and .qmd isn't actually R Markdown I thought it better to add it as a markdown file extension.