Add 2 new aliases to Unix Assembly
Currently, it seems to be impossible (in my testing) to override a file type to be Unix Assembler. This commit should fix that. Unix Assembler uses the AT&T syntax of x86 assembly, and so the first alias is `at&tasm`. The second alias, `unixasm` is just a shortened version of the name for linguist