Less greedy tokenization, fixes segfault
The tokenizer previously matched punctuation and symbols very greedily.
To do so, it required trailing contexts to differentiate from some types
of comments. This trailing context matching led to segfaults.

This commit changes the tokenizer to use less greedy matching of
symbols, and removes the need for trailing context. It has been tuned to
match most common operators, at least where it matters for
classification (checked with cross-validation).

Fixes #5938