Update vendor.yml
What about XML and IML files inside .idea folders for the projects which have used IntelliJ IDEA or similar programs (like Android Studio).  More information:
https://rider-support.jetbrains.com/hc/en-us/articles/207097529-What-is-the-idea-folder-

Example folder with files that are unrelated to project language:
https://user-images.githubusercontent.com/27622683/183767936-43be5dec-490b-4d34-80f4-b1fc67fccb1b.png

These are also accumulated in the project folder and can sometimes be pushed with the rest of the files.