starlark: support recognition of WORKSPACE.bazel
Both WORKSPACE and WORKSPACE.bazel are valid names for the WORKSPACE
file. The latter takes precedent, even though it is an alias because
other projects may have a similarly named file.

Alternatively, *.bazel could be added to extensions, but .bzl is
the recommended extension.

This adds a WORKSPACE.bazel file from
https://github.com/google/skia/blob/main/WORKSPACE.bazel