Replace `\R` in regexps, for portability.
This replaces `\R` with `\r?\n` in two places, which actually slightly changes the meaning, but I believe the regexps were using `\R` because that particular file type is common on Windows.

In a third place, `\R` is replaced with `^`. This works because it follows `(?m:.*?)`, which matches newline characters.