Make regexps in `heuristics.yml` more portable (#6417)
* Escape curly braces in regexps, for portability.

* Replace `\<` with `<` in regexps, for portability.

* Replace `\R` in regexps, for portability.

This replaces `\R` with `\r?\n` in two places, which actually slightly changes the meaning, but I believe the regexps were using `\R` because that particular file type is common on Windows.

In a third place, `\R` is replaced with `^`. This works because it follows `(?m:.*?)`, which matches newline characters.

* Remove some uses of `(?m)` from regexps, for portability.

This flag means something different in Ruby vs. other languages, including JavaScript.

* Change `\N` replacement to `(?:\r?\n|\r)`.

* Change "match any char" patterns to also match `\r`.

This affects three regexps I previously touched to remove `(?m)`. `(?:.|\r)`
matches any character in most languages, but not in JS, so this commit switches
to `(?:.|[\r\n])`.

---------

Co-authored-by: John Gardner <gardnerjohng@gmail.com>
Co-authored-by: Colin Seymour <colin@github.com>