New Centroid-based Classifier (#5103)
* New Centroid-based Classifier

Training:

* A fixed vocabulary is set to all tokens that appear in, at least, 2
  samples.
* All out-of-vocabulary tokens are discarded.
* For every token, we set its Inverse Class Frequency (ICF) to
`log(ct / cf) + 1` where `ct` is the total number of classes and `cf` is
the number of classes where the token occurs.
* Each sample is converted to a vector of `tf * icf` for every token in
the vocabulary. `tf` is `1 + log(freq)`, where `freq` is the
number of occurrences of the token in the given sample.
* Samples are L2-normalized.
* For each class (language), we compute the centroid of all its training
samples by averaging them and L2-normalizing the result.

Classification:

* For a new sample, we get the L2-normalized vector with `tf * icf`
terms for every known token, then classify the sample using the nearest
centroid. Cosine similarity is used as similarity measure for this.

* Fixture file is now detected as Raku

* Update lib/linguist/samples.rb

Co-authored-by: Colin Seymour <colin@github.com>

* Update test/test_classifier.rb

Co-authored-by: Colin Seymour <colin@github.com>

* Add exec bit

* Adjust acceptable errors

* Remove two useless samples

* Add a better R sample

* Remove fixmes

* Remove empty lines

---------

Co-authored-by: Colin Seymour <colin@github.com>
Co-authored-by: Colin Seymour <colin@symr.io>
Co-authored-by: John Gardner <gardnerjohng@gmail.com>