lib/linguist/repository.rb: use Linguist::Source::Repository
Change the 'repository' argument to 'Linguist::Repository' &&
'Linguist::LazyBlob' from a 'Rugged::Repository' to an instance of
'Linguist::Source::Repository'. This will allow users of Linguist to easily
configure and use a custom repository interface.

There are two methods that don't have a clear or useful parallel in a
generic repository interface and are more specific to Rugged: 'read_index'
and 'current_tree'. For both of these methods, raise a 'NotImplementedError'
if any repository instance that's not a 'RuggedRepository' calls them, and
return the legacy value for ones that are 'RuggedRepository'.

Also for backward-compatibility purposes, users can still initialize a
'Linguist::Repository' or 'Linguist::LazyBlob' with a 'Rugged::Repository';
it will be wrapped in the 'Linguist::Source::RuggedRepository' in the
initialization method.

Finally, update 'test_repository.rb' to test both a 'RuggedRepository'
instance and a mocked always-empty repository.

Signed-off-by: Victoria Dye <vdye@github.com>