Refactor `Linguist::Repository` to isolate Rugged usage  (#7094)
* lib/linguist/source: add Repository & Diff interfaces

Add interfaces representing a generic "Repository" and "Diff", containing
functions currently handled by the Rugged repository instance in
'Linguist::Repository'. Inheriting from these interfaces will allow for
alternative implementations of the functions used to traverse and analyze a
repository, e.g. using a different API for Git storage or a different SCM
altogether.

For now, the interfaces are unused.

Signed-off-by: Victoria Dye <vdye@github.com>

* lib/linguist/source: add RuggedRepository

Add Rugged implementations of the 'Repository', 'Diff', and 'Diff::Delta'
interfaces matching existing usage in 'Linguist::Repository' &
'Linguist::LazyBlob'. In a subsequent commit, this will allow us to
substitute an instance of the 'Repository' interface for what is currently
direct usage of a 'Rugged::Repository'.

Signed-off-by: Victoria Dye <vdye@github.com>

* lib/linguist/repository.rb: use Linguist::Source::Repository

Change the 'repository' argument to 'Linguist::Repository' &&
'Linguist::LazyBlob' from a 'Rugged::Repository' to an instance of
'Linguist::Source::Repository'. This will allow users of Linguist to easily
configure and use a custom repository interface.

There are two methods that don't have a clear or useful parallel in a
generic repository interface and are more specific to Rugged: 'read_index'
and 'current_tree'. For both of these methods, raise a 'NotImplementedError'
if any repository instance that's not a 'RuggedRepository' calls them, and
return the legacy value for ones that are 'RuggedRepository'.

Also for backward-compatibility purposes, users can still initialize a
'Linguist::Repository' or 'Linguist::LazyBlob' with a 'Rugged::Repository';
it will be wrapped in the 'Linguist::Source::RuggedRepository' in the
initialization method.

Finally, update 'test_repository.rb' to test both a 'RuggedRepository'
instance and a mocked always-empty repository.

Signed-off-by: Victoria Dye <vdye@github.com>

* lib/linguist/source/rugged.rb: add method_missing delegation

Add a 'method_missing' implementation to 'RuggedRepository' to delegate all
unmatched methods to the internal Rugged objects of each. This is done for
backward-compatibility purposes; users of Linguist can access the
'repository' member of 'Linguist::Repository', so they may rely on
interacting directly with the 'Rugged::Repository'. The 'method_missing'
delegate ensures that those interactions will generally continue to work
(the main exception being explicit type checking performed on
'Linguist::Repository.repository').

Signed-off-by: Victoria Dye <vdye@github.com>

---------

Signed-off-by: Victoria Dye <vdye@github.com>